# perf: Use `ByteString` instead of `[Char]` for input parsing.