@jtdaugherty 2022-03-09
Have you yet attempted a build with `--enable-tests`?