@iphydf 2022-03-09
> Have you yet attempted a build with `--enable-tests`?

No, I'll try that now.